Lesson 5: Custom Views - Added custom compass view to detail fragment.  This compass will align it's needle according to direction data provided in the Wind data string.
However, I was unable to successfully pass layout values from the framework, so this has hard coded values in the meantime to get it working.

Also, there appears to be a new error with FetchWeatherTask, there's an error thrown when trying to hit the weather API.  This is not fixed in this commit.